test(engine): isolate skill recipe validation
