+++ packages/galacean/package.json
@@ -7,7 +7,8 @@
   },
   "license": "MIT",
   "scripts": {
-    "b:types": "tsc"
+    "b:types": "tsc && pnpm run validate:engine-knowledge",
+    "validate:engine-knowledge": "node scripts/validate-engine-knowledge.mjs"
   },
   "types": "types/index.d.ts",
   "main": "dist/main.js",
