+++ packages/galacean/tests/EngineKnowledgeSkill.test.ts
@@ -1,16 +1,7 @@
 import { execFileSync } from "node:child_process";
-import { mkdtempSync, readFileSync, readdirSync, rmSync, writeFileSync } from "node:fs";
-import { tmpdir } from "node:os";
+import { readFileSync, readdirSync } from "node:fs";
 import { dirname, join, relative, resolve, sep } from "node:path";
 import { fileURLToPath } from "node:url";
-import {
-  createProgram,
-  formatDiagnosticsWithColorAndContext,
-  getPreEmitDiagnostics,
-  ModuleKind,
-  ModuleResolutionKind,
-  ScriptTarget
-} from "typescript";
 import { describe, expect, it } from "vitest";
 
 const packageRoot = fileURLToPath(new URL("../", import.meta.url));
@@ -51,6 +42,9 @@ describe("engine-knowledge Skill contract", () => {
       const sourcePath = resolve(skillRoot, source);
       const content = readFileSync(sourcePath, "utf8");
       expect(content.length).toBeGreaterThan(0);
+      expect(content.match(/^```/gm)?.length ?? 0).toBe(
+        [...content.matchAll(/```[^\r\n]*\r?\n[\s\S]*?\r?\n```/g)].length * 2
+      );
       for (const match of content.matchAll(/\]\(([^)#]+\.md)(?:#[^)]+)?\)/g)) {
         if (/^[a-z][a-z\d+.-]*:/i.test(match[1])) {
           continue;
@@ -65,50 +59,6 @@ describe("engine-knowledge Skill contract", () => {
     }
 
     expect([...reachable].sort()).toEqual(markdownFiles);
-  });
-
-  it("type-checks every retained runtime recipe against the packaged declarations", () => {
-    const examples = files(skillRoot)
-      .filter((path) => path.endsWith(".md"))
-      .flatMap((path) => {
-        const content = readFileSync(resolve(skillRoot, path), "utf8");
-        const codeBlocks = [...content.matchAll(/```([^\r\n]*)\r?\n([\s\S]*?)\r?\n```/g)];
-        expect(content.match(/^```/gm)?.length ?? 0).toBe(codeBlocks.length * 2);
-        return codeBlocks.filter((match) => match[1] === "ts").map((match) => match[2]);
-      });
-    expect(examples.length).toBeGreaterThan(0);
-
-    const temporaryRoot = mkdtempSync(join(tmpdir(), "engine-knowledge-recipes-"));
-    try {
-      const files = examples.map((example, index) => {
-        const path = resolve(temporaryRoot, `recipe-${index + 1}.ts`);
-        writeFileSync(path, example);
-        return path;
-      });
-      const program = createProgram(files, {
-        allowSyntheticDefaultImports: true,
-        baseUrl: packageRoot,
-        module: ModuleKind.ESNext,
-        moduleResolution: ModuleResolutionKind.Node10,
-        noEmit: true,
-        paths: { "@galacean/engine": ["types/index.d.ts"] },
-        skipLibCheck: true,
-        strict: true,
-        target: ScriptTarget.ESNext
-      });
-      const diagnostics = getPreEmitDiagnostics(program);
-      if (diagnostics.length > 0) {
-        throw new Error(
-          formatDiagnosticsWithColorAndContext(diagnostics, {
-            getCanonicalFileName: (fileName) => fileName,
-            getCurrentDirectory: () => packageRoot,
-            getNewLine: () => "\n"
-          })
-        );
-      }
-    } finally {
-      rmSync(temporaryRoot, { recursive: true, force: true });
-    }
   });
 
   it("ships the complete Skill in the npm tarball", () => {
